+++ packages/spacecat-shared-data-access/test/it/db.test.js
@@ -14,7 +14,7 @@
 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import dynamoDbLocal from 'dynamo-db-local';
+import DynamoDbLocal from 'dynamo-db-local';
 import Joi from 'joi';
 
 import { isIsoDate } from '@adobe/spacecat-shared-utils';
@@ -107,7 +107,10 @@ describe('DynamoDB Integration Test', async () => {
     process.env.AWS_ACCESS_KEY_ID = 'dummy';
     process.env.AWS_SECRET_ACCESS_KEY = 'dummy';
 
-    dynamoDbLocalProcess = dynamoDbLocal.spawn({ port: 8000, sharedDb: true });
+    dynamoDbLocalProcess = DynamoDbLocal.spawn({
+      port: 8000,
+      sharedDb: true,
+    });
 
     await sleep(5000); // give db time to start up
 
@@ -211,7 +214,7 @@ describe('DynamoDB Integration Test', async () => {
     const sites = await dataAccess.getSitesByOrganizationID(organizations[0].getId());
 
     expect(sites.length).to.be.lessThanOrEqual(Math.trunc(NUMBER_OF_SITES / NUMBER_OF_ORGANIZATIONS)
-        + (NUMBER_OF_SITES % NUMBER_OF_ORGANIZATIONS));
+      + (NUMBER_OF_SITES % NUMBER_OF_ORGANIZATIONS));
 
     sites.forEach((site) => {
       checkSite(site);
@@ -577,5 +580,13 @@ describe('DynamoDB Integration Test', async () => {
 
     updatedSite = await dataAccess.getSiteByID(siteToUpdate.getId());
     expect(updatedSite.getAuditConfig().getAuditTypeConfig('type1').disabled()).to.be.false;
+  });
+
+  it('removes organization', async () => {
+    const organizations = await dataAccess.getOrganizations();
+    const organization = organizations[0];
+    await expect(dataAccess.removeOrganization(organization.getId())).to.eventually.be.fulfilled;
+    const organizationAfterRemoval = await dataAccess.getOrganizationByID(organization.getId());
+    expect(organizationAfterRemoval).to.be.null;
   });
 });